Fixes to the API per Halogen's request.
Memory Box:
- when querying memory box, it does not stop returning me memories.  So
when you reach the bottom of the screen, more memories are loaded (the
same memories keep re-loading). Reported Oct 3.
- I tested and tested and just couldn't get this to happen on any of the
Membox calls (All, Tweets or images). Do you have an example I can look
at?

- removing item from memory box API not working.  Reported Oct 4.
- Fixed

- adding a photo to Memory Box will need to be done on the server side.
We will add the tweet associated with the photo and expect the server to
add the photo to Memory Box.  Similarly, when the user chooses to delete
a photo, we will send the tweet associated with the photo.
- Can you give me an example of the call you would do and result you
expect?

General Server Issues:
- in the next day or two, we will switch over from the staging server to
the Production server to complete testing.  Please let me know if that
will be a problem.
- This will not be a problem, switch anytime.

- the server is not storing subscribed events (My Events). A test
account is still being used to retrieve a list of My Events. These
Events do not correspond to the actual user's Events.

- I've testing this thoroughly and it seems to be working correctly. I
can successfully subscribe to events without a problem. The MyEvents
call returns both subscribed to Events as well as event that user
created.

- I did however find a few problems in the Get requests for Subscribed
Events. I fixed the MyEvents call as well as added a new one that just
returns Subscribed Events.

Subscribed Events
http://epilogger.com/Api/Events/Me/Subscribed/{userId}/{page}/{count}
Will return a list of Subscribed Events for the given UserId. It’s
ordered descending by Subscribed Date.
Return List<ApiEvent>
Sample
Same structure as above

- the server sometimes will timeout when making searches in large
events. The app's timeout has been set to 20 seconds. Searching in a
large event (ie. "new" in "New York Fashion Week") will often cause a
timeout in which case the app will say "Search Error". We can adjust the
timeout delay, but a user would not like to wait much longer than 20
seconds.
- There is nothing that can be done to speed this up at this time. The
error is an acceptable solution.